Separate runner command ownership by compile lane
The runner host now keeps tracing and command dispatch in lib.rs, strategy launch parsing in run.rs, and database/collector operations in an ops-only module. This preserves the default full runner while making lean replay builds compile without the ops command ownership surface.

Constraint: Phase 1 must not alter the default full runner command surface

Constraint: Existing dirty strategy and market-data files are unrelated and remain unstaged

Rejected: Add a new datactl binary in this slice | feature-gated runner-host modules provide the required ownership split with less CLI churn

Confidence: high

Scope-risk: narrow

Directive: Keep future ops commands in ops.rs or a dedicated ops host; do not reintroduce market-data imports at runner-host top level

Tested: cargo check -p new-ploy-runner --no-default-features --features lean-replay

Tested: cargo check -p new-ploy-runner

Tested: cargo run -q -p new-ploy-runner -- --help

Tested: cargo run -q -p new-ploy-runner --no-default-features --features lean-replay -- --help

Tested: cargo run -q -p new-ploy-runner --no-default-features --features lean-replay -- check-db

Tested: scoped rustfmt --check and git diff --check

Not-tested: live database check/quote collection against real DATABASE_URL
